backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
dj-database-url==2.2.0
gunicorn==22.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
